# central-server/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: NEOPRO Central Server API
  description: |
    API RESTful pour la gestion centralisée des écrans publicitaires NEOPRO.

    ## Authentification

    L'API utilise l'authentification JWT. Obtenez un token via `/api/auth/login` et incluez-le dans le header `Authorization: Bearer <token>`.

    Les cookies HttpOnly sécurisés sont également supportés pour les applications web.

    ## MFA (Multi-Factor Authentication)

    Les comptes administrateurs peuvent activer l'authentification à deux facteurs (TOTP).
    Si MFA est activé, le login retournera `requireMfa: true` et nécessitera un code TOTP.

    ## Rate Limiting

    - Auth endpoints: 10 req/15 min
    - API générale: 100 req/min
    - Endpoints sensibles: 30 req/min
    - Uploads vidéo: 10/heure

    ## Codes d'erreur

    - `400` - Requête invalide
    - `401` - Non authentifié
    - `403` - Accès refusé (permissions insuffisantes)
    - `404` - Ressource non trouvée
    - `429` - Trop de requêtes (rate limit)
    - `500` - Erreur serveur
  version: 1.0.0
  contact:
    name: Support NEOPRO
    email: support@neopro.fr
servers:
  - url: https://api.neopro.fr
    description: Production
  - url: http://localhost:3001
    description: Développement
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: neopro_token
  schemas:
...
    Site:
      type: object
      properties:
        id:
          type: string
          format: uuid
        site_name:
          type: string
        club_name:
          type: string
        location:
          type: object
          properties:
            city:
              type: string
            region:
              type: string
            country:
              type: string
            coordinates:
              type: object
              properties:
                lat:
                  type: number
                lng:
                  type: number
        sports:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [online, offline, maintenance, error]
        last_seen_at:
          type: string
          format: date-time
        software_version:
          type: string
        hardware_model:
          type: string
...
    AuditLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        action:
          type: string
        user_id:
          type: string
          format: uuid
        user_email:
          type: string
        target_type:
          type: string
        target_id:
          type: string
        details:
          type: object
        ip_address:
          type: string
        created_at:
          type: string
          format: date-time
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/sites:
    get:
      tags: [Sites]
      summary: Liste des sites
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [online, offline, maintenance, error]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Liste paginée des sites
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Site'
    post:
      tags: [Sites]
      summary: Créer un site
      description: Réservé aux admins et operators
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [site_name, club_name]
              properties:
                site_name:
                  type: string
                club_name:
                  type: string
                location:
                  type: object
                sports:
                  type: array
                  items:
                    type: string
                hardware_model:
                  type: string
      responses:
        '201':
          description: Site créé
          content:
            application/json:
              schema:
                type: object
                properties:
                  site:
                    $ref: '#/components/schemas/Site'
                  apiKey:
                    type: string
                    description: Clé API (affichée une seule fois)
...
  /api/videos:
    get:
      tags: [Videos]
      summary: Liste des vidéos
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Liste des vidéos
    post:
      tags: [Videos]
      summary: Uploader une vidéo
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [video]
              properties:
                video:
                  type: string
                  format: binary
                category:
                  type: string
                subcategory:
                  type: string
                title:
                  type: string
      responses:
        '201':
          description: Vidéo uploadée
...
  /api/groups:
    get:
      tags: [Groups]
      summary: Liste des groupes
      responses:
        '200':
          description: Liste des groupes
    post:
      tags: [Groups]
      summary: Créer un groupe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type]
              properties:
                name:
                  type: string
                description:
                  type: string
                type:
                  type: string
                  enum: [sport, geography, version, custom]
                filters:
                  type: object
      responses:
        '201':
          description: Groupe créé
  /api/audit:
    get:
      tags: [Audit]
      summary: Logs d'audit
      description: Réservé aux admins
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            maximum: 100
        - name: action
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Logs d'audit
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
                  total:
                    type: integer
                  page:
                    type: integer
                  totalPages:
                    type: integer
...